Add typescript
adds typescript, gitignore, and a readme with some project setup stuff.
