+++ .gitignore
@@ -1,0 +1,3 @@
+auth.json
+node_modules
+built
