+++ README.md
@@ -1,0 +1,17 @@
+# Regis
+
+Trivia chatbot with a heart of gold.
+
+## Setup
+
+Copy auth.json.example and name it `auth.json` on root with the actual token value.
+
+## TODOs
+
+* ~~~add typescript~~~
+* get some persistence up in here
+  * add some knex migration up and down scripts for initial db schema
+* make a trivia member sheduler
+  * make the schedule stuff interactive via 
+* flush out some actual trivia game mechanisms
+* have fun!
