+++ auth.json.example
@@ -1,0 +1,3 @@
+{
+    "botUserToken": "replace me with the proper token"
+}
