+++ tsconfig.json
@@ -1,0 +1,17 @@
+{
+    "compilerOptions": {
+        "module": "commonjs",
+        "outDir": "./built",
+        "target": "es6",
+        "lib": [ "es6", "esnext.asynciterable" ],
+        "sourceMap": true,
+        "removeComments": false,
+        "experimentalDecorators": true,
+        "baseUrl": "./src",
+        "rootDir": "./src",
+        "incremental": true
+    },
+    "exclude": [
+        "node_modules"
+    ]
+}
